tests: use autobahn test suite as docker image
According to the documentation from the Autobahn Suite GitHub page, it
seems like they do not support Python 3 which is currently the default
Python version used with our GitHub Actions.

The simplest way to workaround it properly seems to be using the Docker
image that the maintainers of the project suggest to use for running the
suite. This is more future-proof way that we should rely upon.

This commit also removes some redundant steps that are not required.